refactor: 게시글, 댓글 좋아요 개수 반환, 서비스단 폴더구조 변경
