+++ src/main/java/com/example/news_feed/admin/controller/AdminCommentController.java
@@ -1,14 +1,13 @@
 package com.example.news_feed.admin.controller;
 
-import com.example.news_feed.admin.serviceImpl.AdminCommentServiceImpl;
+import com.example.news_feed.admin.service.serviceImpl.AdminCommentServiceImpl;
 import com.example.news_feed.comment.dto.request.UpdateCommentDto;
 import com.example.news_feed.comment.dto.response.CommentDetailDto;
 import com.example.news_feed.comment.dto.response.CommentResponseDto;
-import com.example.news_feed.comment.serviceImpl.CommentServiceImpl;
+import com.example.news_feed.comment.service.serviceImpl.CommentServiceImpl;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
@@ -17,8 +16,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
