+++ src/main/java/com/example/news_feed/follow/controller/FollowApicontroller.java
@@ -2,11 +2,10 @@ package com.example.news_feed.follow.controller;
 
 import com.example.news_feed.follow.dto.response.FollowResponseDto;
 import com.example.news_feed.follow.dto.response.FollowingPostDto;
-import com.example.news_feed.follow.serviceImpl.FollowServiceImpl;
+import com.example.news_feed.follow.service.serviceImpl.FollowServiceImpl;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
@@ -15,8 +14,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
