+++ src/main/java/com/example/news_feed/post/service/serviceImpl/PostLikeServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.news_feed.post.serviceImpl;
+package com.example.news_feed.post.service.serviceImpl;
 
 import com.example.news_feed.post.domain.Post;
 import com.example.news_feed.post.domain.PostLike;
