+++ src/main/java/com/example/news_feed/user/controller/MypageApiController.java
@@ -1,20 +1,17 @@
 package com.example.news_feed.user.controller;
 import com.example.news_feed.comment.dto.response.CommentDetailDto;
 import com.example.news_feed.post.dto.response.PostDetailDto;
-import com.example.news_feed.post.serviceImpl.PostServiceImpl;
-import com.example.news_feed.user.domain.User;
 import com.example.news_feed.user.dto.response.UserDetailDto;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import com.example.news_feed.user.dto.request.PwdUpdateDto;
 import com.example.news_feed.user.dto.request.UserUpdateDto;
 import com.example.news_feed.user.dto.response.UserResponseDto;
-import com.example.news_feed.user.serviceImpl.MyPageServiceImpl;
+import com.example.news_feed.user.service.serviceImpl.MyPageServiceImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
